GroundCheck & PlayerMovement Script
I fixed the groundcheck, updated Grass and Dirt tiles and rule tile, seperated Title controller from the Movfe script
