--- Kings Guard/Assets/Scripts/Components.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3d4926ca9efb77140a4d3fb866d0b5cd
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
