--- Kings Guard/Assets/Scripts/Components/HealthComponent.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using Unity.Entities;
-
-public struct HealthComponent : IComponentData {
-    public float currentHealth;
-    public float maxHealth;
-}
