--- Kings Guard/Assets/Scripts/Components/HealthComponent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bbf56868c29d0e341829826eda9ae6ae
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
