--- Kings Guard/Assets/Scripts/Systems.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e4e4f13ddc9df6742bd01222c377f4f8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
