--- Kings Guard/Assets/Scripts/Systems/DamageSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d9b03962e2a00d0458605f1dc178dc1b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
