+++ Kings Guard/Packages/manifest.json
@@ -1,6 +1,8 @@
 {
   "dependencies": {
     "com.unity.2d.animation": "3.2.2",
+    "com.unity.2d.entities": "0.22.0-preview.8",
+    "com.unity.2d.ik": "2.0.2-preview.1",
     "com.unity.2d.pixel-perfect": "2.0.4",
     "com.unity.2d.psdimporter": "2.1.4",
     "com.unity.2d.sprite": "1.0.0",
