+++ Kings Guard/ProjectSettings/TagManager.asset
@@ -13,9 +13,9 @@ TagManager:
   - UI
   - 
   - 
-  - 
-  - 
-  - 
+  - Ground
+  - Player
+  - Background
   - 
   - 
   -
